requirements.txt
fastapi
uvicorn
gunicorn
python-dotenv
requests
sqlalchemy
psycopg2-binary
python-telegram-bot
